panel: move code from viewlet to composite. ViewletEvent -> CompositeEvent
