+++ src/vs/workbench/parts/files/test/browser/events.test.ts
@@ -8,7 +8,7 @@
 import * as assert from 'assert';
 import URI from 'vs/base/common/uri';
 import {Event, PropertyChangeEvent} from 'vs/base/common/events';
-import {CommandEvent, ViewletEvent, EditorEvent} from 'vs/workbench/common/events';
+import {CommandEvent, EditorEvent} from 'vs/workbench/common/events';
 import {LocalFileChangeEvent} from 'vs/workbench/parts/files/common/files';
 import {FileImportedEvent} from 'vs/workbench/parts/files/browser/fileActions';
