+++ src/vs/workbench/test/browser/events.test.ts
@@ -10,7 +10,7 @@ import URI from 'vs/base/common/uri';
 import * as Paths from 'vs/base/common/paths';
 import * as Files from 'vs/platform/files/common/files';
 import {Event, PropertyChangeEvent} from 'vs/base/common/events';
-import {CommandEvent, ViewletEvent, EditorEvent} from 'vs/workbench/common/events';
+import {CommandEvent, CompositeEvent, EditorEvent} from 'vs/workbench/common/events';
 
 let FileChangesEvent = Files.FileChangesEvent;
 
@@ -98,13 +98,13 @@ suite("Workbench Events", () => {
 		return URI.file(Paths.join('C:\\', path));
 	}
 
-	test("Viewlet Event", function() {
-		let viewletId = "foo.bar";
+	test("Composite Event", function() {
+		let compositeId = "foo.bar";
 		let origEvent = {};
-		let event = new ViewletEvent(viewletId, origEvent);
+		let event = new CompositeEvent(compositeId, origEvent);
 
 		assert.strictEqual(event.originalEvent, origEvent);
-		assert.strictEqual(event.viewletId, viewletId);
+		assert.strictEqual(event.compositeId, compositeId);
 		assert(event.time);
 	});
 });
